Project 01/requirement.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.4
comtypes==1.1.14
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cssselect==1.2.0
cycler==0.11.0
debugpy==1.7.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
docker==6.1.3
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.103.1
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.41.1
fqdn==1.5.1
frozenlist==1.4.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.35
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.57.0
gTTS==2.3.2
h11==0.9.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
ImageHash==4.3.1
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
incremental==22.10.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
mistune==3.0.1
MouseInfo==0.1.3
moviepy==1.0.3
mss==9.0.1
multidict==6.0.4
multimethod==1.9.1
mysql-connector-python==8.0.33
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandas-profiling==3.2.0
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
playsound==1.2.2
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.8.2
pydantic_core==2.4.0
pydeck==0.8.0
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.16.1
Pympler==1.0.1
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywhatkit==5.4
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
Scrapy==2.10.0
seaborn==0.12.2
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sklearn==0.0.post7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
streamlit==1.26.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-serving-api==2.5.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tldextract==3.4.4
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==1.26.16
utils==1.0.1
uvicorn==0.23.2
validators==0.22.0
vboxapi==1.0
ve==0.0.1
virtualenv==20.23.0
visions==0.7.4
w3lib==2.1.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug==2.3.4
wikipedia==1.4.0
wrapt==1.15.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==3.15.0
zope.interface==6.0
